Day 07 part one
This one was really hard! I knew I wanted to use a graph to solve it,
but I didn't think it would be so tough. I started with the sample case
in the problem statement until I got it working. There were a number of
issues.

1. I didn't know `petgraph`, so learning the library was tough.
2. I made some stupid typos and was scrolled up in the sample problem
   file. PEBCAK :(
3. I had some flawed attempts at calculating the totals. Eventually
   I realized that actual filesystems have to walk the directories to
   find the size so I just built my graph then calculated the sizes.
   Getting the final data structure for the HashMap keys was the last
   thing, since some directory names were repeated.

This solution won't work if directories are `ls`'d more than once, but
that case can be fixed by checking if a node exists when we `cd`.