Dec 08 part one
This was pretty straight forward. The first approach I tried only
checked the immediate neighbors. I had to modify it to look all the way
to the edges in every direction, but that was fairly straight forward.

Really happy with how quick this came together! No real issues with Rust
except for dealing with char.to_digit(10) creating a u32 instead of
usize.