Dec 08 part two
Straight up copy and paste then modification. I think it would be
possible to use the same function to get each direction and then do
a check outside the function, but w/e.

Ran into some issues with checking the edges (which are always 0). Also
had a stupid copy and paste error where I was subtracting the x instead
of the y. Figured it out in the end :)